introduce useFont hook
