+++ src/theme/index.js
@@ -1,4 +1,5 @@
 export { default as defaultDarkTheme } from "./darkTheme";
 export { default as defaultLightTheme } from "./lightTheme";
 export { default as useTheme } from "./useTheme";
+export { default as useFont } from "./useFont";
 export * from "./helpers";
