+++ src/theme/useTheme.js
@@ -2,7 +2,7 @@ import { useMemo, useLayoutEffect } from "react";
 import deepmerge from "deepmerge";
 import lightTheme from "./lightTheme";
 
-const useTheme = (themeOverrides, fontConfig) => {
+const useTheme = (themeOverrides) => {
   const res = useMemo(
     () =>
       lightTheme && themeOverrides
@@ -12,45 +12,11 @@ const useTheme = (themeOverrides, fontConfig) => {
   );
 
   useLayoutEffect(() => {
-    if (fontConfig) {
-      applyFonyAssets(fontConfig, res);
-    }
     Object.keys(res).forEach((key) => {
       document.documentElement.style.setProperty(`--${key}`, res[key]);
     });
-  }, [themeOverrides, fontConfig]);
+  }, [themeOverrides]);
   return res;
-};
-
-const applyFonyAssets = (fontConfig, theme) => {
-  var newStyle = document.createElement("style");
-  newStyle.appendChild(
-    document.createTextNode(`
-@font-face {
-  font-family: ${fontConfig.fontFamilyText};
-  src: url(${fontConfig.semiBoldUrl})
-    format("${fontConfig.format}");
-  font-weight: ${theme["font-weight-semi-bold"]};
-  font-style: normal;
-}
-@font-face {
-  font-family: ${fontConfig.fontFamilyText};
-  src: url(${fontConfig.lightUrl})
-    format("${fontConfig.format}");
-  font-weight: ${theme["font-weight-light"]};
-  font-style: normal;
-}
-@font-face {
-  font-family: ${fontConfig.fontFamilyText};
-  src: url(${fontConfig.regularUrl})
-    format("${fontConfig.format}");
-  font-weight: ${theme["font-weight-regular"]};
-  font-style: normal;
-}
-`)
-  );
-  newStyle.type = "text/css";
-  document.head.appendChild(newStyle);
 };
 
 export default useTheme;
